Tried to use cache for excel and rewrote the loops
Tried to use cache for excel and rewrote the loops taking them out of the if statements because, duh, of course that would be more inefficient to have them inside ifs. Also added code to connect to database and populate it from the excel so I can just work with the database on the next update.